// src/model/userInfo.model.ts
export interface IFormUserInfo {
  allowReceiveExcel: string;
  allowReceivePdf: string;
  birthday: string;
  createdBy: string;
  createdDate: string;
  customerId: string;
  deptId: string;
  deptName: string;
  email: string;
  employeeCode: string;
  fullName: string;
  id: number;
  idNumber: string;
  ipAddress: string;
  isViettel: string;
  issueBy: string;
  issueDate: string;
  lastModifiedBy: string;
  lastModifiedDate: string;
  passwordRaw: string;
  phone: string;
  roles: any[];
  status: number;
  username: string;
}